a lot of bugs fixed. now should be ready for training.

+++ trunk/anji_2_01-choe/src/com/anji/neatpacman/ghosts/GhostsPlayer.java
@@ -27,9 +27,9 @@ public class GhostsPlayer implements Player
     activator.reset();
   }
 
-  public int[] move(GameState[] ghostsStates)
+  public byte[] move(GameState[] ghostsStates)
   {
-    int[] rst = new int[ghostsStates.length];
+    byte[] rst = new byte[ghostsStates.length];
     
     for (int i = 0; i < ghostsStates.length; ++i)
     {
